checkt nu of de afbeelding leeg is zo ja dan upload ie niks

+++ functions/veranderHomepage.php
@@ -11,14 +11,15 @@ function cmsHomepage() {
 				<tr><td>
 					<form method="post" action="" enctype="multipart/form-data">
 						<input type="file" value="Upload afbeelding" name="afbeelding1">
-						<input type="submit" class="btnVeranderHomepage" value="Upload afbeelding" name="uploadafbeelding1">
+						<input type="submit" class="btnVeranderHomepage" value="Upload afbeelding1" name="uploadafbeelding1">
 					</form>
 				</td>
 				
-				<td>
+				</tr>
+				<tr><td>
 					<form method="post" action="" enctype="multipart/form-data">
 						<input type="file" value="Upload afbeelding" name="afbeelding2">
-						<input type="submit" class="btnVeranderHomepage" value="Upload afbeelding" name="uploadafbeelding2">
+						<input type="submit" class="btnVeranderHomepage" value="Upload afbeelding2" name="uploadafbeelding2">
 					</form>
 				</td></tr>
 				
@@ -58,49 +59,57 @@ function tekstVeranderHomepage() {
 }
 
 function veranderAfbeelding1() {
+	if (!$_FILES['afbeelding1']['name'] == "") {
 
-	$files = glob('images/homepage/afbeelding1/*');
-	foreach ($files as $file) {
-		if (is_file($file))
-			unlink($file);
-	}
-
-	$photo = $_FILES['afbeelding1'];
-	if (isset($_POST['uploadafbeelding1'])) {
-		var_dump($photo);
-		if (!is_uploaded_file($photo['name'])) {
-			move_uploaded_file($photo['tmp_name'], "images/homepage/afbeelding1/" . $photo['name']);
-
-			$query = 'UPDATE  `content`
-		SET content="images/homepage/afbeelding1/' . $photo['name'] . '" WHERE id="2"';
-			$resultaat = mysql_query($query);
-
-		} else {
-			echo 'Failed';
+		$files = glob('images/homepage/afbeelding1/*');
+		foreach ($files as $file) {
+			if (is_file($file))
+				unlink($file);
 		}
+
+		$photo = $_FILES['afbeelding1'];
+		if (isset($_POST['uploadafbeelding1'])) {
+			if (!is_uploaded_file($photo['name'])) {
+				move_uploaded_file($photo['tmp_name'], "images/homepage/afbeelding1/" . $photo['name']);
+
+				$query = 'UPDATE  `content`
+		SET content="images/homepage/afbeelding1/' . $photo['name'] . '" WHERE id="2"';
+				$resultaat = mysql_query($query);
+
+			} else {
+				echo 'Failed';
+			}
+		}
+	} else {
+		$output = "Afbeelding is leeg";
+		return $output;
 	}
 }
 
 function veranderAfbeelding2() {
+	if (!$_FILES['afbeelding2']['name'] == "") {
 
-	$files = glob('images/homepage/afbeelding2/*');
-	foreach ($files as $file) {
-		if (is_file($file))
-			unlink($file);
-	}
-
-	$photo = $_FILES['afbeelding2'];
-	if (isset($_POST['uploadafbeelding2'])) {
-		var_dump($photo);
-		if (!is_uploaded_file($photo['name'])) {
-			move_uploaded_file($photo['tmp_name'], "images/homepage/afbeelding2/" . $photo['name']);
-
-			$query = 'UPDATE  `content`
-		SET content="images/homepage/afbeelding2/' . $photo['name'] . '" WHERE id="3"';
-			$resultaat = mysql_query($query);
-
-		} else {
-			echo 'Failed';
+		$files = glob('images/homepage/afbeelding2/*');
+		foreach ($files as $file) {
+			if (is_file($file))
+				unlink($file);
 		}
+
+		$photo = $_FILES['afbeelding2'];
+		if (isset($_POST['uploadafbeelding2'])) {
+			if (!is_uploaded_file($photo['name'])) {
+				move_uploaded_file($photo['tmp_name'], "images/homepage/afbeelding2/" . $photo['name']);
+
+				$query = 'UPDATE  `content`
+		SET content="images/homepage/afbeelding2/' . $photo['name'] . '" WHERE id="3"';
+				$resultaat = mysql_query($query);
+
+			} else {
+				echo 'Failed';
+			}
+		}
+	} else {
+		$output = "Afbeelding is leeg";
+		return $output;
 	}
 }
